Rewrite game::handlers::handle to use the new Ring struct for inputs

+++ common/src/player.rs
@@ -201,7 +201,7 @@ impl PlayerState {
 
 #[derive(Debug, Default, Clone, Copy, Serialize, Deserialize)]
 pub struct PlayerInput {
-    pub target_tick: u16,
+    pub target_tick: u64,
     pub forward: bool,
     pub backward: bool,
     pub left: bool,
